+++ source/models/seasons/Participation.js
@@ -42,6 +42,10 @@ exports.initModel = function (sequelize) {
 			type: DataTypes.INTEGER,
 			defaultValue: 0
 		},
+		toastsRaised: {
+			type: DataTypes.INTEGER,
+			defaultValue: 0
+		},
 		dqCount: {
 			type: DataTypes.INTEGER,
 			defaultValue: 0